code editing finished

+++ src/app/note-components/components/note-thumbnail/note-thumbnail.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { Router } from '@angular/router';
-import { Note } from 'src/app/core/models/note.model';
+import { Note } from 'src/app/core/models/note.interface';
 import { tap } from 'rxjs'; 
 import { NoteService } from 'src/app/core/services/note.service';
 
@@ -37,15 +37,17 @@ export class NoteThumbnailComponent {
 
   onDeleteNote(event?: any) {
     event.stopPropagation();
-    this.noteService.deleteNote(this.note.id).pipe(
-      tap(res => {
-        if(res) {
-          this.isDeleted = true;
-          if(this.router.url === `/notes/${this.note.id}`)
-            this.router.navigateByUrl('');
-        }
-      })
-    ).subscribe();
+    if(this.note.id) {
+      this.noteService.deleteNote(this.note.id).pipe(
+        tap(res => {
+          if(res) {
+            this.isDeleted = true;
+            if(this.router.url === `/notes/${this.note.id}`)
+              this.router.navigateByUrl('');
+          }
+        })
+      ).subscribe();
+    }
   }
 
 }
